Вносит правки после созвона
