+++ js/api.js
@@ -13,19 +13,16 @@ const ErrorText = {
   SEND_DATA: 'Не удалось отправить форму. Попробуйте ещё раз',
 };
 
-/*
-const load = (route, errorText, method = Method.GET, body = null) =>
-  fetch(`${BASE_URL}${route}`, {method, body})
-    .then((response) => {
-      if (!response.ok) {
-        throw new Error();
-      }
-      return response.json();
-    })
-    .catch(() => {
-      throw new Error(errorText);
-    });
-*/
+const showDataError = () => {
+  const errorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
+  const errorMessage = errorTemplate.cloneNode(true);
+
+  document.body.append(errorMessage);
+
+  setTimeout(() => {
+    errorMessage.remove();
+  }, 5000);
+};
 
 const load = async (route, errorText, method = Method.GET, body = null) => {
   try {
@@ -45,4 +42,4 @@ const getData = () => load(Route.GET_DATA, ErrorText.GET_DATA);
 
 const sendData = (body) => load(Route.SEND_DATA, ErrorText.SEND_DATA, Method.POST, body);
 
-export {getData, sendData};
+export { getData, sendData, showDataError };
