+++ js/utils.js
@@ -1,0 +1,3 @@
+const isEscapeKey = (evt) => evt.key === 'Escape';
+
+export { isEscapeKey };
